Better code structure
