+++ nbproject/private/private.xml
@@ -3,8 +3,7 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/H:/EquationConverter/EquationConverter/lenguajes/src/lenguajes/TAC.java</file>
-            <file>file:/H:/EquationConverter/EquationConverter/lenguajes/src/lenguajes/Lenguajes.java</file>
+            <file>file:/D:/EquationConverter/EquationConverter/lenguajes/src/lenguajes/Lenguajes.java</file>
         </group>
     </open-files>
 </project-private>
